requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-adk
google-generativeai
python-dotenv
pydantic